Move Configuration to a separate module
